requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
blinker==1.8.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.78
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.32
Mako==1.3.10
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.25.2
openai==1.109.1
orjson==3.10.15
packaging==24.2
pandas==2.3.3
propcache==0.2.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sentence_transformers==5.1.1
faiss_cpu==1.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
